+++ plattera-ui/src/app/restaurants/menu/demoData.ts
@@ -1,0 +1,144 @@
+import { MenuItem } from "./types";
+
+export const demoMenuItems: MenuItem[] = [
+  {
+    itemId: 1,
+    itemName: "Butter Chicken",
+    restId: 0,
+    itemPrice: "12.99",
+    itemDescription:
+      "Tender chicken pieces in a rich, creamy tomato-based curry sauce",
+    itemSweet: false,
+    itemSpicy: true,
+    itemSweetLevel: 2,
+    itemSpiceLevel: 4,
+    itemAvailable: true,
+    itemBestSeller: true,
+    itemIsVeg: false,
+    itemIsJain: false,
+    itemImage: "/DummyDishImage.jpg",
+    category: "non-veg",
+  },
+  {
+    itemId: 2,
+    itemName: "Paneer Tikka",
+    restId: 0,
+    itemPrice: "9.99",
+    itemDescription:
+      "Grilled cottage cheese cubes marinated in spices and yogurt",
+    itemSweet: false,
+    itemSpicy: true,
+    itemSweetLevel: 1,
+    itemSpiceLevel: 3,
+    itemAvailable: true,
+    itemBestSeller: true,
+    itemIsVeg: true,
+    itemIsJain: false,
+    itemImage: "/DummyDishImage.jpg",
+    category: "veg",
+  },
+  {
+    itemId: 3,
+    itemName: "Gulab Jamun",
+    restId: 0,
+    itemPrice: "5.99",
+    itemDescription:
+      "Sweet milk-solid balls soaked in rose-flavored sugar syrup",
+    itemSweet: true,
+    itemSpicy: false,
+    itemSweetLevel: 5,
+    itemSpiceLevel: 0,
+    itemAvailable: true,
+    itemBestSeller: false,
+    itemIsVeg: true,
+    itemIsJain: true,
+    itemImage: "/DummyDishImage.jpg",
+    category: "desserts",
+  },
+  {
+    itemId: 4,
+    itemName: "Veg Biryani",
+    restId: 0,
+    itemPrice: "11.99",
+    itemDescription:
+      "Fragrant basmati rice cooked with mixed vegetables and aromatic spices",
+    itemSweet: false,
+    itemSpicy: true,
+    itemSweetLevel: 1,
+    itemSpiceLevel: 3,
+    itemAvailable: false,
+    itemBestSeller: true,
+    itemIsVeg: true,
+    itemIsJain: false,
+    itemImage: "/DummyDishImage.jpg",
+    category: "veg",
+  },
+  {
+    itemId: 5,
+    itemName: "Chicken Tikka Masala",
+    restId: 0,
+    itemPrice: "13.99",
+    itemDescription: "Grilled chicken in a creamy, spiced tomato-based curry",
+    itemSweet: false,
+    itemSpicy: true,
+    itemSweetLevel: 2,
+    itemSpiceLevel: 4,
+    itemAvailable: true,
+    itemBestSeller: true,
+    itemIsVeg: false,
+    itemIsJain: false,
+    itemImage: "/DummyDishImage.jpg",
+    category: "non-veg",
+  },
+  {
+    itemId: 6,
+    itemName: "Mango Lassi",
+    restId: 0,
+    itemPrice: "4.99",
+    itemDescription: "Sweet yogurt drink with mango pulp",
+    itemSweet: true,
+    itemSpicy: false,
+    itemSweetLevel: 4,
+    itemSpiceLevel: 0,
+    itemAvailable: true,
+    itemBestSeller: false,
+    itemIsVeg: true,
+    itemIsJain: true,
+    itemImage: "/DummyDishImage.jpg",
+    category: "drinks",
+  },
+  {
+    itemId: 7,
+    itemName: "Masala Chai",
+    restId: 0,
+    itemPrice: "3.99",
+    itemDescription: "Traditional spiced Indian tea with milk",
+    itemSweet: true,
+    itemSpicy: true,
+    itemSweetLevel: 3,
+    itemSpiceLevel: 2,
+    itemAvailable: true,
+    itemBestSeller: true,
+    itemIsVeg: true,
+    itemIsJain: false,
+    itemImage: "/DummyDishImage.jpg",
+    category: "drinks",
+  },
+  {
+    itemId: 8,
+    itemName: "Rasmalai",
+    restId: 0,
+    itemPrice: "6.99",
+    itemDescription: "Soft cottage cheese dumplings in sweetened milk",
+    itemSweet: true,
+    itemSpicy: false,
+    itemSweetLevel: 5,
+    itemSpiceLevel: 0,
+    itemAvailable: true,
+    itemBestSeller: false,
+    itemIsVeg: true,
+    itemIsJain: true,
+    itemImage: "/DummyDishImage.jpg",
+    category: "desserts",
+  },
+];
